Refactor: move blueprints to directory, routes -> misc, buckets -> unlinked
