+++ flask/app/blueprints/families.py
@@ -3,10 +3,11 @@ from dataclasses import asdict
 from flask import abort, jsonify, request, Blueprint, current_app as app
 from flask_login import current_user, login_required
 from sqlalchemy.orm import joinedload
-from .extensions import db
-from . import models
-from .schemas import FamilySchema
-from .utils import (
+
+from .. import models
+from ..extensions import db
+from ..schemas import FamilySchema
+from ..utils import (
     check_admin,
     filter_datasets_by_user_groups,
     get_current_user,
