+++ flask/app/blueprints/genes.py
@@ -1,11 +1,13 @@
 from dataclasses import asdict
 from typing import Any, Dict
+
 from flask import abort, jsonify, request, Blueprint
 from flask_login import login_required
 from sqlalchemy import func
 from sqlalchemy.orm import contains_eager, joinedload
-from .models import Gene, GeneAlias
-from .utils import csv_response, expects_csv, expects_json, paged, paginated_response
+
+from ..models import Gene, GeneAlias
+from ..utils import csv_response, expects_csv, expects_json, paged, paginated_response
 
 genes_blueprint = Blueprint(
     "genes",
