+++ flask/app/blueprints/tissue_samples.py
@@ -2,11 +2,12 @@ from dataclasses import asdict
 
 from flask import abort, jsonify, request, Blueprint, current_app as app
 from flask_login import current_user, login_required
-from .extensions import db
-from . import models
+
+from .. import models
+from ..extensions import db
 from sqlalchemy.orm import contains_eager, joinedload
-from .schemas import TissueSampleSchema
-from .utils import (
+from ..schemas import TissueSampleSchema
+from ..utils import (
     check_admin,
     filter_datasets_by_user_groups,
     get_current_user,
