+++ flask/app/blueprints/unlinked.py
@@ -3,8 +3,8 @@ from flask_login import current_user, login_required
 from sqlalchemy.orm import joinedload
 from sqlalchemy import or_
 
-from . import models
-from .utils import get_minio_client
+from .. import models
+from ..utils import get_minio_client
 
 
 bucket_blueprint = Blueprint(
