+++ flask/app/blueprints/variants.py
@@ -8,9 +8,9 @@ from sqlalchemy import distinct, func
 from sqlalchemy.orm import aliased, contains_eager
 from sqlalchemy.sql import and_, or_
 
-from . import models
-from .extensions import db
-from .utils import (
+from .. import models
+from ..extensions import db
+from ..utils import (
     expects_csv,
     expects_json,
     filter_datasets_by_user_groups,
